requirements.txt
git+https://github.com/huggingface/accelerate.git
git+https://github.com/huggingface/diffusers.git
git+https://github.com/huggingface/peft.git
transformers==4.57.3
huggingface_hub
sentencepiece
torchvision
safetensors
runpod
boto3
requests
Pillow
numpy
torch
